handle --offline

+++ src/photo-display.py
@@ -15,23 +15,20 @@ def OpenImage(image, resolution):
 
 def main(args):
     settings = Settings(args.config)    
-
-    immich = ImmichConnection(settings.ImmichServerUrl, settings.ApiKey)
-    album_data = []
-    for album_id in settings.Albums:
-        album = immich.sync_album(album_id)
-        if album is not None:
-            album_data.append(album)
-
     screen = Screen(settings)
 
     if not args.no_screen:
         screen.init_inky()
 
-    database = ImageDatabase(settings, screen.resolution)    
-    database.process_albums(immich)
-    database.purge_missing(immich)
-        
+    immich = ImmichConnection(settings.ImmichServerUrl, settings.ApiKey)
+    database = ImageDatabase(settings, screen.resolution)
+
+    if not args.offline:
+        for album_id in settings.Albums:
+            immich.sync_album(album_id)
+        database.process_albums(immich)
+        database.purge_missing(immich)
+      
     while True:            
         target_image = database.get_random_image()
         if target_image is not None:          
@@ -48,6 +45,7 @@ def main(args):
     
 parser = argparse.ArgumentParser(description='Display photos from an Immich album onto an Inky Screen.')
 parser.add_argument('--config', help='config json file', default="./config.json", required=False)
+parser.add_argument('--offline', help='dont connect to the network and use what is in the database', action='store_true', required=False)
 parser.add_argument('--no-screen', help='dont init inky screen. useful for debugging', action='store_true', required=False)
 if __name__ == "__main__":
     args = parser.parse_args()
